Remediate NPM Advisory 1164
After running `npm install`, npm warns of a severe security
vulnerabilty: https://www.npmjs.com/advisories/1164.

This change to `package.json` ensures the version of handlebars that is
installed (because of jest) is unaffected.

This can be removed once jest releases a newer version that requires an
unaffected version of handlebars.