Fix the write tests
I misinterpreted something from Sather and thought that the high bit
had to be set on the data for writing to happen at all.  This is not
true.  Instead, there is a flux transition every time the high bit is
set as the data is left-shifted out of the data register.  The
erroneous test has been removed.

At the same time, I finally understand what `skip` does and documented
that.