Add tests for the Disk II card (#133)
* Add tests for the DiskII card

This change adds basic read tests for nibble-based disks for the
DiskII card and fixes a few minor errors.

These tests are in preparation for refactoring.

* Add write tests

These are some basic tests of writing to nibble disks.  In the process,
one minor bug was found, fixed and documented.

* Fix the write tests

I misinterpreted something from Sather and thought that the high bit
had to be set on the data for writing to happen at all.  This is not
true.  Instead, there is a flux transition every time the high bit is
set as the data is left-shifted out of the data register.  The
erroneous test has been removed.

At the same time, I finally understand what `skip` does and documented
that.

* Add tests for saving and restoring Disk II state

These are not exhaustive tests, but they ensure that some basic state
is saved and restored.